+++ test/transform/resource/before/DataConfiguration.java
@@ -1,4 +1,6 @@
 //CONF: lombok.anyConstructor.suppressConstructorProperties = true
+//CONF: lombok.toString.doNotUseGetters = true
+//CONF: lombok.equalsAndHashCode.doNotUseGetters = true
 @lombok.Data
 class DataConfiguration {
 	final int x;